requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Document Parsing (Phase 2)
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-pptx==0.6.23
chardet==6.0.0

# NLP & Knowledge Extraction (Phase 3)
spacy==3.7.2

# Knowledge Graph (Phase 3) - Graph Database
neo4j==5.16.0

# QA & LLM Integration (Phase 4)
langchain==0.1.0
langchain-openai==0.0.5
openai>=1.10.0,<2.0.0
faiss-cpu==1.13.2
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
werkzeug==3.0.1

# Testing
pytest==7.4.3
pytest-flask==1.2.0